:art: directory 구조 개선
page 디렉토리를 제거하고 해당 페이지 컴포넌트들을 /components로 옮기고 파일을 index.tsx로 변경

페이지에서만 쓰이는 컴포넌트들은 모두 해당 디렉토리로 이동
공통적으로 쓰이는 컴포넌트들은 ./components/common으로 이동
