--- src/components/Modal/NewAlarmModal.tsx
@@ -1,51 +1,0 @@
-/** @jsx jsx */
-import { ReactElement, useEffect } from 'react';
-
-import { css, jsx } from '@emotion/react';
-import styled from '@emotion/styled';
-import classnames from 'classnames';
-
-import { COLOR } from '../../constant/color';
-import { COMMON } from '../../constant/message';
-import Modal from './Modal';
-
-interface Props {
-  closeHandler: () => void;
-  className?: string;
-}
-
-const InnerModalStyle = css`
-  width: 100%;
-  height: auto;
-  padding: 2rem 0;
-`;
-
-const ContentsWrapper = styled.div`
-  text-align: center;
-  color: ${COLOR.TEXT_BLACK};
-  letter-spacing: -0.03rem;
-`;
-
-function NewAlarmModal({ closeHandler, className }: Props): ReactElement {
-  useEffect(() => {
-    const closeModal = setTimeout(() => {
-      closeHandler();
-    }, 1200);
-    return () => {
-      clearTimeout(closeModal);
-    };
-  }, [closeHandler]);
-
-  return (
-    <Modal
-      className={classnames('new-alarm-modal', className)}
-      innerModalStyle={InnerModalStyle}
-    >
-      <ContentsWrapper className="new-alarm-modal__contents body3">
-        {COMMON.NEW_ALARM_MODAL.TEXT}
-      </ContentsWrapper>
-    </Modal>
-  );
-}
-
-export default NewAlarmModal;
