+++ src/components/common/DescriptionItem.tsx
